+++ R/backend-.R
@@ -167,6 +167,7 @@ base_scalar <- sql_translator(
   as.character = sql_cast("TEXT"),
   as.logical = sql_cast("BOOLEAN"),
   as.Date = sql_cast("DATE"),
+  as.POSIXct = sql_cast("TIMESTAMP"),
   # MS SQL - https://docs.microsoft.com/en-us/sql/t-sql/data-types/int-bigint-smallint-and-tinyint-transact-sql
   # Hive - https://cwiki.apache.org/confluence/display/Hive/LanguageManual+Types#LanguageManualTypes-IntegralTypes(TINYINT,SMALLINT,INT/INTEGER,BIGINT)
   # Postgres - https://www.postgresql.org/docs/8.4/static/datatype-numeric.html
@@ -184,6 +185,26 @@ base_scalar <- sql_translator(
   pmax = sql_prefix("GREATEST"),
 
   `%>%` = `%>%`,
+
+  # lubridate ---------------------------------------------------------------
+  # https://en.wikibooks.org/wiki/SQL_Dialects_Reference/Functions_and_expressions/Date_and_time_functions
+  as_date = sql_cast("DATE"),
+  as_datetime = sql_cast("TIMESTAMP"),
+
+  today = function() sql_expr(CURRENT_DATE),
+  now = function() sql_expr(CURRENT_TIMESTAMP),
+
+  # https://modern-sql.com/feature/extract
+  year = function(x) sql_expr(EXTRACT(year %from% !!x)),
+  month = function(x) sql_expr(EXTRACT(month %from% !!x)),
+  day = function(x) sql_expr(EXTRACT(day %from% !!x)),
+  mday = function(x) sql_expr(EXTRACT(day %from% !!x)),
+  yday = sql_not_supported("yday()"),
+  qday = sql_not_supported("qday()"),
+  wday = sql_not_supported("wday()"),
+  hour = function(x) sql_expr(EXTRACT(hour %from% !!x)),
+  minute = function(x) sql_expr(EXTRACT(minute %from% !!x)),
+  second = function(x) sql_expr(EXTRACT(second %from% !!x)),
 
   # String functions ------------------------------------------------------
   # SQL Syntax reference links: